AI NewsLetter/pyproject.toml
[tool.poetry]
name = "ai-news-letter"
version = "0.1.0"
description = "AI News Letter"
authors = ["Anantha, Narayanan"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
crewai = "^0.11.0"
duckduckgo-search = "^4.4.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"